requirements.txt
absl-py==2.1.0
altgraph==0.17.4
annotated-types==0.7.0
anyio==4.9.0
astunparse==1.6.3
attrs==25.3.0
bleach==6.2.0
blinker==1.9.0
certifi==2025.1.31
cffi==2.0.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==46.0.3
cvzone==1.6.1
cycler==0.12.1
dnspython==2.8.0
fastapi==0.115.12
filelock==3.18.0
Flask==3.1.0
flask-cors==6.0.1
Flask-Login==0.6.3
flatbuffers==25.2.10
fonttools==4.56.0
fsspec==2025.3.2
gast==0.6.0
gitdb==4.0.12
GitPython==3.1.44
google-generativeai==0.8.3
google-pasta==0.2.0
greenlet==3.2.4
grpcio==1.71.0
h11==0.14.0
h5py==3.13.0
idna==3.10
itsdangerous==2.2.0
jax==0.5.3
jaxlib==0.5.3
Jinja2==3.1.6
joblib==1.5.0
kaggle==1.7.4.2
kagglehub==0.3.11
keras==3.9.0
keyboard==0.13.5
kiwisolver==1.4.8
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mdurl==0.1.2
mediapipe==0.10.21
ml_dtypes==0.5.1
mpmath==1.3.0
mysql-connector-python==9.4.0
namex==0.0.8
networkx==3.4.2
numpy==1.26.4
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
opt_einsum==3.4.0
optree==0.14.1
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pefile==2023.2.7
pillow==11.1.0
protobuf==4.25.6
psutil==7.0.0
py-cpuinfo==9.0.0
pyarrow==19.0.1
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pyFirmata==1.1.0
pygame==2.6.1
Pygments==2.19.1
pyinstaller==6.12.0
pyinstaller-hooks-contrib==2025.1
pylance==0.24.1
pymongo==4.15.3
pyparsing==3.2.1
pyserial==3.5
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.1
pywin32-ctypes==0.2.3
PyYAML==6.0.2
requests==2.32.3
rich==13.9.4
scikit-learn==1.7.2
scipy==1.15.2
seaborn==0.13.2
selenium==4.30.0
sentencepiece==0.2.0
setuptools==76.0.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.1
SQLAlchemy==2.0.43
starlette==0.46.2
sympy==1.13.1
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorflow==2.19.0
tensorflow-hub==0.16.1
termcolor==2.5.0
text-unidecode==1.3
tf_keras==2.19.0
thop==0.1.1.post2209072238
threadpoolctl==3.6.0
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
trio==0.29.0
trio-websocket==0.12.2
typing-inspection==0.4.0
typing_extensions==4.12.2
tzdata==2025.1
ultralytics==8.3.105
ultralytics-thop==2.0.14
urllib3==2.3.0
uvicorn==0.35.0
webencodings==0.5.1
websocket-client==1.8.0
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.2
wsproto==1.2.0
python-dotenv==1.0.1
elevenlabs==1.59.0
